+++ demo/src/components/handsontable/basic-demo.ts
@@ -1,13 +1,13 @@
 /* tslint:disable:no-any no-magic-numbers prefer-function-over-method */
 import { Component } from '@angular/core';
-import { genData } from './data';
+import { getBasicData } from './data';
 
 @Component({
   selector: 'basic-demo',
   template: require('./basic-demo.html')
 })
 export class BasicDemoComponent {
-  private data: any[] = genData(10);
+  private data: any[] = getBasicData(10);
   private colHeaders: string[] = ['ID', 'First Name', 'Last Name', 'Address',
     'Favorite food', 'Price', 'Is active'];
   private columns: any[] = [
@@ -28,16 +28,12 @@ export class BasicDemoComponent {
     },
     {
       data: 'product.description',
-      source: 'product.options',
-      optionField: 'description',
-      type: 'autocomplete',
-      strict: false,
-      visibleRows: 4
+      renderer: 'text'
     },
     {
       data: 'price',
       type: 'numeric',
-      format: '$ 0,0.00'
+      numericFormat: { pattern: '$0,0.00', culture: 'en-US' }
     },
     {
       data: 'isActive',
@@ -56,10 +52,12 @@ export class BasicDemoComponent {
   };
 
   private afterChange(e: any) {
+    // tslint:disable-next-line:no-console
     console.log(e);
   }
 
   private afterOnCellMouseDown(e: any) {
+    // tslint:disable-next-line:no-console
     console.log(e);
   }
 }