+++ demo/src/typings.d.ts
@@ -1,17 +1,11 @@
-/// <reference path="../../scripts/handsontable.d.ts" />
-
 // Typings reference file, you can add your own global typings here
 // https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
-// tslint:disable
-declare const System: any;
-declare const ENV:string;
-// google code-prettify
-declare const PR:any;
-
 declare const chroma: any;
 
 declare module jasmine {
-  interface Matchers {
+  interface Matchers<T> {
     toHaveCssClass(expected: any): boolean;
   }
 }
+
+declare const require: (module: string) => any;